Delete unused code, rename some stuff
