--- packages/esm-service-queues-app/src/active-visits/active-visits-row-actions.scss
@@ -1,4 +1,0 @@
-.transitionMenuContainer {
-  display: inline;
-  vertical-align: top;
-}
