+++ packages/esm-service-queues-app/src/clear-queue-entries-modal/clear-queue-entries.resource.ts
@@ -1,4 +1,4 @@
-import { endPatientStatus } from '../active-visits/active-visits-table.resource';
+import { endPatientStatus } from '../service-queues.resource';
 import { type QueueEntry } from '../types';
 
 export async function batchClearQueueEntries(queueEntries: Array<QueueEntry>) {
