+++ packages/esm-service-queues-app/src/queue-patient-linelists/queue-services-table.component.tsx
@@ -1,7 +1,7 @@
 import React, { useMemo } from 'react';
 import { useTranslation } from 'react-i18next';
 import QueuePatientBaseTable from './queue-linelist-base-table.component';
-import { useServiceQueueEntries } from '../active-visits/active-visits-table.resource';
+import { useServiceQueueEntries } from '../service-queues.resource';
 
 const ServicesTable: React.FC = () => {
   const { t } = useTranslation();
