+++ packages/esm-service-queues-app/src/routes.json
@@ -59,12 +59,6 @@
       "component": "pastVisitSummary",
       "slot": "previous-visit-summary-slot"
     },
-
-    {
-      "name": "active-visits-row-actions",
-      "component": "activeVisitsRowActions",
-      "slot": "queue-table-serve-patient-slot"
-    },
     {
       "name": "visit-form-queue-fields",
       "component": "visitFormQueueFields",
